Added modality in visibility
